Concluidos os EndPoints Principais

+++ Models/Users.cs
@@ -7,8 +7,8 @@
         public string Email { get; set; }
         public string HashedPassword { get; set; }
         public DateTime? DateOfBirth { get; set; }
-        public double? Weight { get; set; }
-        public double? GoalWeight { get; set; }
+        public float? Weight { get; set; }
+        public float? GoalWeight { get; set; }
         public string ActivityLevel { get; set; }
         public DateTime DateJoined { get; set; }
         public DateTime? LastLogin { get; set; }
